move code for return checks to ReturnCheckHelper
